[diffusion] VDN-H3: review pass; per-tensor fp8 as the pipeline default
Fix the chunk-boundary scan grid under anchor_frames="both" (the branch
renumbers inner frames from 1, so chunk boundaries sit at frame_offset 1;
the gather had read zero window-complement states) and test the branch
against the plain frame chain and the eager kernel chain through the real
forward. int64 addressing in the branch Triton kernels (the fused-qkv view
overflows int32 at the paper clip on one GPU). The per-head scalars ride the
async exchange; the gate's `up` lives on VDNOutputGate; the hybrid core is
split into sized helpers; the window plan constructor is split.

fp8: Fp8Config(per_tensor_online) set from PipelineConfig.online_fp8_per_tensor
(VDN-H3 only) replaces the env var; the per-tensor path goes through
apply_fp8_linear, falls back per layer for unaligned or fp32 layers, and the
prequantized input is a LinearMethodBase hook. Fewer tests, each guarding a
distinct failure mode; the 2-GPU exchange test moves to the registered
standalone suite; the fused SwiGLU kernel gets a can_use predicate and a
registered kernel test. Narrative comments removed.

+++ python/sglang/kernels/ops/diffusion/attention/vdn_linear_branch_triton.py
@@ -69,7 +69,7 @@ def _tconv_act_kernel(
         r = rows + dt - 2
         ok = valid & (r >= 0) & (r < T)  # zero padding, both ends
         v = tl.load(
-            X + (r[:, None] * S_ + pid_s) * C_ + chan[None, :],
+            X + (r[:, None].to(tl.int64) * S_ + pid_s) * C_ + chan[None, :],
             mask=ok[:, None],
             other=0.0,
         ).to(tl.float32)
@@ -82,11 +82,11 @@ def _tconv_act_kernel(
         y = y * inv[:, None]
     if FRAME_MAJOR:
         # [T, HEADS, S, D]: the readout bmm reads this layout directly
-        dst = ((rows[:, None] * HEADS + pid_h) * S_ + pid_s) * D_ + tl.arange(0, D_)[
-            None, :
-        ]
+        dst = (
+            (rows[:, None].to(tl.int64) * HEADS + pid_h) * S_ + pid_s
+        ) * D_ + tl.arange(0, D_)[None, :]
     else:
-        dst = (rows[:, None] * S_ + pid_s) * C_ + chan[None, :]
+        dst = (rows[:, None].to(tl.int64) * S_ + pid_s) * C_ + chan[None, :]
     tl.store(OUT + dst, y.to(OUT.dtype.element_ty), mask=valid[:, None])
 
 
@@ -156,7 +156,7 @@ def _silu_l2norm_kernel(
     valid = rows < N
     offs = tl.arange(0, D_)
     x = tl.load(
-        X + rows[:, None] * stride_n + pid_h * stride_h + offs[None, :],
+        X + rows[:, None].to(tl.int64) * stride_n + pid_h * stride_h + offs[None, :],
         mask=valid[:, None],
         other=0.0,
     ).to(tl.float32)
@@ -168,9 +168,11 @@ def _silu_l2norm_kernel(
         # row n = frame * S_ + s -> [F, H, S_, D]
         frame = rows // S_
         pos = rows - frame * S_
-        dst = ((frame[:, None] * H + pid_h) * S_ + pos[:, None]) * D_ + offs[None, :]
+        dst = (
+            (frame[:, None].to(tl.int64) * H + pid_h) * S_ + pos[:, None]
+        ) * D_ + offs[None, :]
     else:
-        dst = (rows[:, None] * H + pid_h) * D_ + offs[None, :]
+        dst = (rows[:, None].to(tl.int64) * H + pid_h) * D_ + offs[None, :]
     tl.store(OUT + dst, y.to(OUT.dtype.element_ty), mask=valid[:, None])
 
 
@@ -235,7 +237,7 @@ def _frame_stats_prep_kernel(
     valid = s < S_
     offs = tl.arange(0, D_)
     rows = f * S_ + s  # token rows
-    src = (rows[:, None] * H + h) * D_ + offs[None, :]  # [F*S, H, d]
+    src = (rows[:, None].to(tl.int64) * H + h) * D_ + offs[None, :]  # [F*S, H, d]
     dst = ((f * H + h) * S_ + s)[:, None] * D_ + offs[None, :]  # [F, H, S, d]
     k = tl.load(K + src, mask=valid[:, None], other=0.0)
     v = tl.load(V + src, mask=valid[:, None], other=0.0)
@@ -318,7 +320,7 @@ def _linear_epilogue_kernel(
     offs = tl.arange(0, D_)
     src = ((f * H + h) * S_ + s)[:, None] * D_ + offs[None, :]
     rows = f * S_ + s
-    dst = (rows[:, None] * H + h) * D_ + offs[None, :]
+    dst = (rows[:, None].to(tl.int64) * H + h) * D_ + offs[None, :]
     r = tl.load(R + src, mask=valid[:, None], other=0.0).to(tl.float32)
     ms = tl.sum(r * r, axis=1) / D_
     w = tl.load(W + offs).to(tl.float32)
